+++ PinedaAppBE/PinedaApp/Controllers/ExperienceController.cs
@@ -1,6 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using PinedaApp.Configurations;
 using PinedaApp.Contracts;
+using PinedaApp.Models;
 using PinedaApp.Models.Errors;
 using PinedaApp.Services;
 
@@ -24,7 +26,7 @@ namespace PinedaApp.Controllers
             }
             catch (PinedaAppException ex)
             {
-                ErrorResponse error = new ErrorResponse(ex.Message);
+                ErrorResponse error = new(ex.Message);
                 return StatusCode(ex.ErrorCode, error);
             }
         }
@@ -39,11 +41,12 @@ namespace PinedaApp.Controllers
             }
             catch (PinedaAppException ex)
             {
-                ErrorResponse error = new ErrorResponse(ex.Message);
+                ErrorResponse error = new(ex.Message);
                 return StatusCode(ex.ErrorCode, error);
             }
         }
 
+        [Authorize]
         [HttpPost]
         public IActionResult CreateExperience(ExperienceRequest request)
         {
@@ -61,22 +64,30 @@ namespace PinedaApp.Controllers
             {
                 if (ex.InnerException is ValidationException validationEx)
                 {
-                    ErrorResponse response = new ErrorResponse(validationEx.ValidationErrors.Errors);
+                    ErrorResponse response = new(validationEx.ValidationErrors.Errors);
                     return BadRequest(response);
                 }
                 else
                 {
-                    ErrorResponse response = new ErrorResponse(ex.Message);
+                    ErrorResponse response = new(ex.Message);
                     return StatusCode(ex.ErrorCode, response);
                 }
             }
         }
 
+        [Authorize]
         [HttpPut("{id:int}")]
         public IActionResult UpdateExperience(int id, ExperienceRequest request)
         {
             try
             {
+                int userId = GetUserId();
+                if (userId == 0 || !CheckUserOwner(id, userId))
+                {
+                    ErrorResponse forbidden = new("Not Allowed to Update Data");
+                    return StatusCode(403, forbidden);
+                }
+
                 ExperienceResponse response = _experience.UpsertExperience(request, id);
                 if (response.Id == id) return NoContent();
 
@@ -92,28 +103,36 @@ namespace PinedaApp.Controllers
             {
                 if (ex.InnerException is ValidationException validationEx)
                 {
-                    ErrorResponse response = new ErrorResponse(validationEx.ValidationErrors.Errors);
+                    ErrorResponse response = new(validationEx.ValidationErrors.Errors);
                     return BadRequest(response);
                 }
                 else
                 {
-                    ErrorResponse response = new ErrorResponse(ex.Message);
+                    ErrorResponse response = new(ex.Message);
                     return StatusCode(ex.ErrorCode, response);
                 }
             }
         }
 
+        [Authorize]
         [HttpDelete("{id:int}")]
         public IActionResult DeleteExperience(int id)
         {
             try
             {
+                int userId = GetUserId();
+                if (userId == 0 || !CheckUserOwner(id, userId))
+                {
+                    ErrorResponse forbidden = new("Not Allowed to Delete Data");
+                    return StatusCode(403, forbidden);
+                }
+
                 _experience.DeleteExperience(id);
                 return NoContent();
             }
             catch (PinedaAppException ex)
             {
-                ErrorResponse error = new ErrorResponse(ex.Message);
+                ErrorResponse error = new(ex.Message);
                 return StatusCode(ex.ErrorCode, error);
             }
         }
