+++ PinedaAppBE/PinedaApp/Controllers/PortfolioController.cs
@@ -1,14 +1,17 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using PinedaApp.Configurations;
 using PinedaApp.Contracts;
 using PinedaApp.Models.Errors;
 using PinedaApp.Services;
+using System.Security.Claims;
 
 namespace PinedaApp.Controllers
 {
     public class PortfolioController : BaseApiController
     {
         private readonly IPortfolioService _portfolioService;
+        private readonly int _userId = 0;
         public PortfolioController(IPortfolioService portfolioService)
         {
             _portfolioService = portfolioService;
@@ -24,7 +27,7 @@ namespace PinedaApp.Controllers
             }
             catch (PinedaAppException ex)
             {
-                ErrorResponse error = new ErrorResponse(ex.Message);
+                ErrorResponse error = new(ex.Message);
                 return StatusCode(ex.ErrorCode, error);
             }
         }
@@ -39,11 +42,12 @@ namespace PinedaApp.Controllers
             }
             catch (PinedaAppException ex)
             {
-                ErrorResponse error = new ErrorResponse(ex.Message);
+                ErrorResponse error = new(ex.Message);
                 return StatusCode(ex.ErrorCode, error);
             }
         }
 
+        [Authorize]
         [HttpPost]
         public IActionResult CreatePortfolio(PortfolioRequest request)
         {
@@ -61,22 +65,29 @@ namespace PinedaApp.Controllers
             {
                 if (ex.InnerException is ValidationException validationEx)
                 {
-                    ErrorResponse response = new ErrorResponse(validationEx.ValidationErrors.Errors);
+                    ErrorResponse response = new(validationEx.ValidationErrors.Errors);
                     return BadRequest(response);
                 }
                 else
                 {
-                    ErrorResponse response = new ErrorResponse(ex.Message);
+                    ErrorResponse response = new(ex.Message);
                     return StatusCode(ex.ErrorCode, response);
                 }
             }
         }
 
+        [Authorize]
         [HttpPut("{id:int}")]
         public IActionResult UpdatePortfolio(PortfolioRequest request, int id)
         {
             try
             {
+                int userId = GetUserId();
+                if (userId == 0 || !CheckUserOwner(id, userId))
+                {
+                    ErrorResponse forbidden = new("Not Allowed to Delete Data");
+                    return StatusCode(403, forbidden);
+                }
                 PortfolioResponse response = _portfolioService.UpsertPortfolio(request, id);
 
                 if (response.Id == id) return NoContent();
@@ -92,28 +103,36 @@ namespace PinedaApp.Controllers
             {
                 if (ex.InnerException is ValidationException validationEx)
                 {
-                    ErrorResponse response = new ErrorResponse(validationEx.ValidationErrors.Errors);
+                    ErrorResponse response = new(validationEx.ValidationErrors.Errors);
                     return BadRequest(response);
                 }
                 else
                 {
-                    ErrorResponse response = new ErrorResponse(ex.Message);
+                    ErrorResponse response = new(ex.Message);
                     return StatusCode(ex.ErrorCode, response);
                 }
             }
         }
 
+        [Authorize]
         [HttpDelete]
         public IActionResult DeletePortfolio(int id)
         {
             try
             {
+                int userId = GetUserId();
+                if (userId == 0 || !CheckUserOwner(id, userId))
+                {
+                    ErrorResponse forbidden = new("Not Allowed to Delete Data");
+                    return StatusCode(403, forbidden);
+                }
+
                 _portfolioService.DeletePortfolio(id);
                 return NoContent();
             }
             catch (PinedaAppException ex)
             {
-                ErrorResponse response = new ErrorResponse(ex.Message);
+                ErrorResponse response = new(ex.Message);
                 return StatusCode(ex.ErrorCode, response);
             }
         }
